requirements.txt
asgiref==3.5.2
Django==4.1.1
django-extensions==3.2.0
django-filter==22.1
gunicorn==20.1.0
sqlparse==0.4.2
whitenoise==6.2.0
